ajaxでいいねをjavascriptを用いて作成
