+++ app/views/articles/show.html.haml
@@ -33,16 +33,11 @@
     -# もしサインインしていたら
     .article_heart.hidden.active-heart
       -# クラスのhiddenをscssで表示を消す用事してクリックされたら表示するようにする
-      = link_to article_like_path(@article), data: {method: 'delete' } do
-        -# link_toでリンクに飛ぶ、article_like_pathの@articleの記事に、data: {method: 'delete' } doでデータを削除したら
-        = image_tag 'heart-active.svg'
-        -# 記事をいいねを押したらheart-active.svgの表示
+      = image_tag 'heart-active.svg'
+      -# heart-active.svgの表示
     .article_heart.hidden.inactive-heart
-      -# クラスのhiddenをscssで表示を消す用事してクリックされたら表示するようにする
-      = link_to article_like_path(@article), data: { method: 'post' } do
-        -# link_toでリンクに飛ぶ、article_like_pathの@articleの記事に、data: {method: 'post' } doでデータをpostをしたら
-        = image_tag 'heart.svg'
-        -# heart.svgの表示
+      = image_tag 'heart.svg'
+      -# heart.svgの表示
 
 .article
   %h2 コメント一覧
@@ -57,3 +52,6 @@
     .btn-secondary
       コメントを追加
       -# link_toでリンクの作成、new_article_comment_pathはcommentsのURL、(@article)はarticleIDの指定。
+
+= javascript_pack_tag 'article'
+-# javascriptのpackの中のarticleを読み込める
