+++ package.json
@@ -9,6 +9,7 @@
     "@rails/webpacker": "4.3.0",
     "axios": "^0.20.0",
     "jquery": "^3.5.1",
+    "rails-ujs": "^5.2.4-4",
     "trix": "^1.2.0",
     "turbolinks": "^5.2.0"
   },
